Don't automatically escape all non-ASCII characters.
This looses a bit of generality in that we are producing a unicode
string, that can't really be turned uinto ASCII form.  On the other hand,
it is not clear that there is much use to support this, while
non-ASCII characters are quite common in modern usage, and it is not
nice if you have to look at them in encoded form.