Closing tag parser should skip trailing whitespace
According to http://www.w3.org/TR/xml/#sec-starttags , end tags are
allowed to have whitespace between the closing tag name and the '>'
character.  The grammar shows:

  ETag ::= '</' Name S? '>'

but the end-tag case does not handle the S? part (the opening tag
seems to handle it, though).

I have tested this diff with

<a>blah</a
><b>asdf</b>

As an example.  Without the change, extraneous CDATA elements
containing "\n>" are interspersed in the resulting document.  After
this change, the above contains no extraneous CDATA (but does still
have a node for the ending newline).